efactor: modularize flow editor functionality with custom hooks
