+++ apps/frontend/src/app/(main)/workflows/[id]/editor/_components/flow-editor.css
@@ -64,20 +64,27 @@
 }
 
 .react-flow__minimap-mask {
-  fill: var(--muted-foreground);
-  opacity: 0.1;
+  fill: var(--muted-foreground) !important;
+  opacity: 0.1 !important;
 }
 
+/* Default node color */
 .react-flow__minimap-node {
-  fill: var(--primary);
+  fill: var(--primary-foreground) !important;
   stroke: var(--background);
   stroke-width: 2px;
+}
+
+/* Selected node color */
+.react-flow__node.selected + .react-flow__minimap-node {
+  fill: var(--primary) !important;
 }
 
 /* Selection styling */
 .react-flow__node.selected {
   border-color: var(--primary) !important;
-  box-shadow: 0 0 0 4px rgba(99, 102, 241, 0.3) !important;
+  box-shadow: 0 0 0 8px rgba(99, 102, 241, 0.3) !important;
+  border-radius: 0.375rem !important;
 }
 
 .react-flow__edge.selected .react-flow__edge-path {
